Update README.md
I'm using this amazing tool for a client-side project, and I do not use require.js. I wanted to see how well this pure JS project would work in a more traditional client-side setting. Works like a champ! I thought it would be handy to update the README.md to reflect this configuration.